requirements.txt
requests
pyyaml
python-dotenv
uvicorn
fastapi
google-cloud-storage
playwright
gspread
google-auth
